requirements.txt
asgiref==3.6.0
boto3==1.26.74
botocore==1.29.74
certifi==2024.12.14
charset-normalizer==3.4.1
da-vinci==0.4.0
Django==4.1.7
django-admin-thumbnails==0.2.9
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.10
jmespath==1.0.1
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
s3transfer==0.6.1
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
typing-extensions==4.12.2
urllib3==1.26.14
